feat(dx12): 完善DX12图形适配器的头文件包含和类型转换
- 统一添加 DirectX 和 WRL 相关头文件包含
- 优化类型转换，确保宽高数据精度一致
- 清理冗余或重复的头文件引用
- 修复视口和裁剪矩形设置中的类型不匹配问题
- 标准化 ComPtr 的使用方式
- 移除不再使用的 DXGI 和 D3D 头文件引用
- 更新 Fence 和 PipelineState 类型定义以提高兼容性
- 确保所有 DX12 相关类正确引入所需依赖项

+++ src/engine/graphic/adapters/dx12/DX12CommandBuffer.cpp
@@ -1,11 +1,14 @@
+#include <directx/d3d12.h>
+#include <directx/d3dx12.h>
+
 #include "DX12CommandBuffer.h"
 #include "DX12RenderDevice.h"
+#include <wrl/client.h>
 #include "DX12Texture.h"
 #include "DX12Buffer.h"
 #include "DX12PipelineState.h"
 #include "DX12Sampler.h"
 
-#include <directx/d3dx12.h>
 #include <sstream>
 #include <algorithm>
 
@@ -137,14 +140,14 @@ void DX12CommandBuffer::BeginRenderPass(const RenderPassDesc& desc) {
         // 设置视口和裁剪矩形
         D3D12_VIEWPORT viewport = {};
         viewport.Width = static_cast<float>(dx12RenderTarget->GetWidth());
-        viewport.Height = dx12RenderTarget->GetHeight();
+        viewport.Height = static_cast<float>(dx12RenderTarget->GetHeight());
         viewport.MinDepth = 0.0f;
         viewport.MaxDepth = 1.0f;
         m_commandList->RSSetViewports(1, &viewport);
 
         D3D12_RECT scissor = {};
-        scissor.right = dx12RenderTarget->GetWidth();
-        scissor.bottom = dx12RenderTarget->GetHeight();
+        scissor.right = static_cast<LONG>(dx12RenderTarget->GetWidth());
+        scissor.bottom = static_cast<LONG>(dx12RenderTarget->GetHeight());
         m_commandList->RSSetScissorRects(1, &scissor);
     }
